Build: test on JDK 8, 9, 10 & 11
This change makes the project build do 4 runs using JDK 8, 9, 10 & 11 intead of running on default which is JDK 11.
After the latest merge the TravisCI build shows red. The best guess is that between last time the PR was built –
an unreasonable time of 3 months ago – and it was merged TravisCI has changed defaults for JDK-based project to
use JDK 11 instead of JDK 8. To prevent similar build failures the project now sets a range of reasonable to
support JDK versions explicitely.